+++ src/amun/gpu/decoder/enc_out_buffer.cu
@@ -25,9 +25,9 @@ EncOutPtr EncOutBuffer::Get()
   return buffer_.remove();
 }
 
-void EncOutBuffer::Get(size_t num, std::vector<BufferOutput> &ret)
+void EncOutBuffer::Get(unsigned num, std::vector<BufferOutput> &ret)
 {
-  for (size_t currNum = 0; currNum < num; ++currNum) {
+  for (unsigned currNum = 0; currNum < num; ++currNum) {
     if (unfinishedEncOutPtr_.get() == nullptr) {
       unfinishedEncOutPtr_ = Get();
       unfinishedInd_ = 0;
